Client-EVM ready for usage
